% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.small-data.R
\docType{data}
\name{l.small}
\alias{l.small}
\title{Espresso truth table with 4 inputs and 3 outputs}
\format{R data frame table}
\usage{
data(l.small)
}
\description{
Espresso compatible truth table generated 
from espresso format file small.esp.
}
\examples{
\dontrun{
# steps to recreate
inpath <- system.file("extdata/espresso/small.esp", package="LogicOpt")
l.small <- logicopt(esp_file=inpath,mode="echo")[1]
}
}
\keyword{Espresso}
\keyword{truth-table}

