% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicopt.R
\name{num_input_values}
\alias{num_input_values}
\title{Find size of input values}
\usage{
num_input_values(tt, n_in)
}
\arguments{
\item{tt}{a truth table where first n_in columns are inputs}

\item{n_in}{number of inputs}
}
\value{
returns a vector of number of unique input values for each input
}
\description{
Find number of unique input values for each input in
tt.
}

