% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicopt.R
\name{print_primes_tt}
\alias{print_primes_tt}
\title{Print the Primes from logicopt(mode="primes")}
\usage{
print_primes_tt(primes, eqn = FALSE, n_in, n_out)
}
\arguments{
\item{primes}{An R data frame table generated by logicopt(...,mode="primes").}

\item{eqn}{Print in equation format.  Default is FALSE.}

\item{n_in}{Number of inputs in the primes truth table.}

\item{n_out}{Number of outputs in the primes truth table.}
}
\value{
None
}
\description{
This function prints the results from logicopt(..,,mode="primes")
option.
}
\examples{
data(l.small)
ptt <- logicopt(l.small,n_in=4,n_out=3,find_dc=TRUE,mode="primes")
print_primes_tt(ptt,eqn=TRUE,4,3) 

}

